+++ jest.config.js
@@ -9,6 +9,7 @@ module.exports = {
 		"<rootDir>/packages/plugin-iobroker/src",
 		"<rootDir>/packages/plugin-lerna/src",
 		"<rootDir>/packages/plugin-license/src",
+		"<rootDir>/packages/plugin-manual-review/src",
 		"<rootDir>/packages/plugin-package/src",
 		"<rootDir>/packages/plugin-version/src",
 		// Add others as necessary
@@ -25,6 +26,8 @@ module.exports = {
 			"<rootDir>/packages/plugin-iobroker/src$1",
 		"^@alcalzone/release-script-plugin-lerna(.*)": "<rootDir>/packages/plugin-lerna/src$1",
 		"^@alcalzone/release-script-plugin-license(.*)": "<rootDir>/packages/plugin-license/src$1",
+		"^@alcalzone/release-script-plugin-manual-review(.*)":
+			"<rootDir>/packages/plugin-manual-review/src$1",
 		"^@alcalzone/release-script-plugin-package(.*)": "<rootDir>/packages/plugin-package/src$1",
 		"^@alcalzone/release-script-plugin-version(.*)": "<rootDir>/packages/plugin-version/src$1",
 		"^@alcalzone/release-script-testing(.*)": "<rootDir>/packages/testing/src$1",